犀牛鸟 #54: MoT ablation boundary tests, routing analysis & dtype fix
- tests/test_mot.py: 27→36 boundary tests (fp16 stability, hard routing
  gradient flow, routing determinism, top_k=E dense routing, scene-aware
  router, train/eval consistency, invalid top_k validation, multi-size
  localconv, shift_size=0 mode)
- scripts/mot_routing_analysis.py: MoTBlock expert routing analysis
  script (router bias extraction, per-layer expert preference, scene-
  conditioned activation patterns, temperature annealing tracking)
- block.py:220/232: fix dtype mismatch in _blend_experts
  (Half×Float→add .to(out.dtype) cast)

+++ ultralytics/nn/modules/mot/block.py
@@ -217,7 +217,7 @@ class MoTBlock(nn.Module):
                         f"→ output {tuple(expert_out.shape)}. All experts must preserve "
                         f"the input tensor shape."
                     )
-                out[batch_idx] = out[batch_idx] + expert_out * w
+                out[batch_idx] = out[batch_idx] + (expert_out * w).to(out.dtype)
             self._last_dispatch_stats = {"mode": "sample_sparse", "expert_calls": expert_calls, "selected_samples": B}
         else:
             for e_idx, expert in enumerate(self.experts):
@@ -229,7 +229,7 @@ class MoTBlock(nn.Module):
                         f"→ output {tuple(expert_out.shape)}. All experts must preserve "
                         f"the input tensor shape."
                     )
-                out = out + expert_out * w
+                out = out + (expert_out * w).to(out.dtype)
             self._last_dispatch_stats = {"mode": "dense", "expert_calls": len(self.experts), "selected_samples": B}
         return out
